BDC-151: Catalog resolution for common views
The catalog of the view object as well as the base table(s) in the view can be
in one of the two forms:
1) Explicitly specified as prodhive, testiceberg etc
2) Explicitly specified as hive or iceberg
3) Not specified.

In case 1, it is sufficient to record the catalog the way it is specified.
In case 2, hive/iceberg resolve to 'prodhive' on prod systems and to
'testhive' on test systems. So, these need to be resolved and stored
correctly as metacat catalogs.
In case 3, the session catalog is used to resolve the objects. Session
catalog can be of type 1 or 2 above and needs to be handled accordingly.